refactor: kotlin 정규 표현식 사용
