+++ src/main/kotlin/edu/example/kotlindevelop/domain/member/controller/AdminController.kt
@@ -1,7 +1,8 @@
-package edu.example.kotlindevelop.member.controller
+package edu.example.kotlindevelop.domain.member.controller
 
-import edu.example.kotlindevelop.member.dto.MemberDTO
-import edu.example.kotlindevelop.member.service.MemberService
+
+import edu.example.kotlindevelop.domain.member.dto.MemberDTO
+import edu.example.kotlindevelop.domain.member.service.MemberService
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.PageRequest
 import org.springframework.data.domain.Pageable
@@ -26,7 +27,7 @@ class AdminController (
         val pageable: Pageable = PageRequest.of(page, pageSize)
         val responseDto: Page<MemberDTO.Response> = memberService.readAll(pageable)
 
-        return ResponseEntity.ok<Page<MemberDTO.Response>>(responseDto)
+        return ResponseEntity.ok(responseDto)
     }
 
     // 아래는 ProductService merge 후 완성
